replace types to dedicated file
