+++ pallets/dca/src/types.rs
@@ -1,4 +1,61 @@
+use codec::{Decode, Encode, MaxEncodedLen};
+use scale_info::TypeInfo;
+use sp_runtime::traits::ConstU32;
+use sp_runtime::BoundedVec;
 pub type Balance = u128;
 pub type AssetId = u32;
 pub type BlockNumber = u64;
 pub type ScheduleId = u32;
+
+const MAX_NUMBER_OF_TRADES: u32 = 5;
+pub const MAX_NUMBER_OF_SCHEDULES_PER_BLOCK: u32 = 20; //TODO: use config for this
+
+#[derive(Encode, Decode, Debug, Eq, PartialEq, Clone, TypeInfo, MaxEncodedLen)]
+pub enum Recurrence {
+	Fixed(u128),
+	Perpetual,
+}
+
+#[derive(Encode, Decode, Debug, Eq, PartialEq, Clone, TypeInfo, MaxEncodedLen)]
+pub enum Order<AssetId> {
+	Sell {
+		asset_in: AssetId,
+		asset_out: AssetId,
+		amount_in: Balance,
+		min_limit: Balance,
+		route: BoundedVec<Trade, ConstU32<MAX_NUMBER_OF_TRADES>>,
+	},
+	Buy {
+		asset_in: AssetId,
+		asset_out: AssetId,
+		amount_out: Balance,
+		max_limit: Balance,
+		route: BoundedVec<Trade, ConstU32<MAX_NUMBER_OF_TRADES>>,
+	},
+}
+
+#[derive(Encode, Decode, Debug, Eq, PartialEq, Clone, TypeInfo, MaxEncodedLen)]
+pub struct Schedule<AssetId, BlockNumber> {
+	pub period: BlockNumber,
+	pub recurrence: Recurrence,
+	pub order: Order<AssetId>,
+}
+
+///A single trade for buy/sell, describing the asset pair and the pool type in which the trade is executed
+#[derive(Encode, Decode, Debug, Eq, PartialEq, Clone, TypeInfo, MaxEncodedLen)]
+pub struct Trade {
+	pub pool: PoolType, //TODO: consider using the same type as in route executor
+	pub asset_in: AssetId,
+	pub asset_out: AssetId,
+}
+
+#[derive(Encode, Decode, Clone, Copy, Debug, Eq, PartialEq, TypeInfo, MaxEncodedLen)]
+pub enum PoolType {
+	Omnipool,
+}
+
+#[derive(Encode, Decode, Debug, Eq, PartialEq, Clone, TypeInfo, MaxEncodedLen)]
+pub struct Bond<AssetId> {
+	pub asset: AssetId,
+	pub amount: Balance,
+}
